Cut the post way down, go with conclusions earlier.
The earlier draft was working through the issues, with more examples quote> to
help myself along. Once I decided how it all ought to work, it made more sense
to remove as much of the stuff I used to arrive at the solution and just
present the problems, more simply, followed by the proposed solutions. I think
this will be much more useful for readers, many of whom won't care about the
process by which I came to these conclusions, but want to know about the
conclusions themselves.